requirements.txt
accelerate==1.6.0
diffusers==0.33.1
transformers==4.46.2
gradio==5.25.2
sentencepiece==0.2.0
pillow==11.1.0
av==12.1.0
numpy==1.26.2
scipy==1.12.0
requests==2.31.0
torchsde==0.2.6
jinja2>=3.1.2
einops
opencv-contrib-python
safetensors
peft
decord
tqdm
imageio==2.31.1
imageio-ffmpeg==0.4.8
ffmpeg-python==0.2.0
# for toolbox
psutil
realesrgan
basicsr-fixed

